Customer functionalities part 1

+++ src/co/nextix/jardine/database/tables/picklists/PCustTypeTable.java
@@ -120,7 +120,7 @@ public class PCustTypeTable {
 		return rowsDeleted;
 	}
 
-	public PicklistRecord getById(int ID) {
+	public PicklistRecord getById(long ID) {
 		PicklistRecord record = null;
 		String MY_QUERY = "SELECT * FROM " + mDatabaseTable + " WHERE "
 				+ KEY_CUSTOMER_TYPE_ROWID + "=?";
